requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.9
Django==3.2.5
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-multi-captcha-admin==1.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
flake8==3.9.2
idna==3.2
inflection==0.5.1
ipython==7.25.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mysqlclient==2.0.3
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
uWSGI==2.0.19.1
wcwidth==0.2.5
